Don't stop looking for the view controller when it's a UINavigationController
I needed to put a UITablBarController inside a UINavigationController (I know, NOT recommended by Apple, but I had to do it nonetheless).

This fixes the stylesheet not being able to be set because when it found a UINavigationController as the root_view_controller, it just gave up. I'm asking the method to keep searching till it gets an actual UIViewController and not blindly trust that the visibleViewController is a UIViewController.